safari-534.53-branch/LayoutTests/sputnik/Conformance/07_Lexical_Conventions/7.3_Line_Terminators/S7.3_A3.2_T3-expected.txt

CONSOLE MESSAGE: line 78: SyntaxError: Parse error
S7.3_A3.2_T3

PASS Expected parsing failure

TEST COMPLETE
